# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы расчета резинотехнических изделий",
        "title": "UltraRezina API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calculations/{calculationId}/materials/{materialId}": {
            "put": {
                "description": "Обновляет параметры связи между расчетом и материалом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-calculations"
                ],
                "summary": "Обновить связь расчета и материала",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID расчета",
                        "name": "calculationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID материала",
                        "name": "materialId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMaterialCalculationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Связь обновлена",
                        "schema": {
                            "$ref": "#/definitions/models.MaterialCalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Расчет не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет материал из конкретного расчета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-calculations"
                ],
                "summary": "Удалить материал из расчета",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID расчета",
                        "name": "calculationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID материала",
                        "name": "materialId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Материал удален из расчета",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Расчет не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.MaterialCalculationResponse": {
            "type": "object",
            "properties": {
                "calculation_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "material": {
                    "$ref": "#/definitions/models.MaterialResponse"
                },
                "material_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "result_freq": {
                    "type": "number"
                },
                "result_percent": {
                    "type": "number"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "models.MaterialResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "density": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "material": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "thickness": {
                    "type": "number"
                }
            }
        },
...
        "models.UpdateMaterialCalculationRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "quantity": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен для аутентификации. Используйте формат: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}